+++ R/embed_url.R
@@ -1,18 +1,16 @@
-#' Embed a video based on URL
+#' Embed video based on URL
 #'
 #' You can use this function to embed video using only the URL and you do not
-#' need any customization beyond the start-time.
-#' It should work for all the services supported by the \code{\link{embed}}
-#' family of functions.
+#' need any customization beyond the start-time. It works for all the
+#' services supported by the [embed()] family of functions.
 #'
-#' This function calls \code{\link{suggest_embed}} then parses and evaluates the code.
-#' If you need to customize the iframe, \code{\link{suggest_embed}} may be more useful to you.
+#' This function calls [suggest_embed()] then parses and evaluates the code.
 #'
-#' @param url character, URL of webpage for video
+#' @param url `character`, URL of web-page for video
 #'
-#' @return An embed object that prints an \code{htmltools::\link[htmltools]{tags}$iframe} element
+#' @inherit embed return
 #'
-#' @seealso \code{\link{suggest_embed}}
+#' @seealso [suggest_embed()]
 #' @examples
 #' embed_url("https://youtu.be/1-vcErOPofQ?t=28s")
 #' @export
@@ -26,4 +24,33 @@ embed_url <- function(url){
   suggest_embed_pure(url) %>%
     parse_text() %>%
     eval()
+}
+
+#' @rdname embed
+#' @export
+#
+embed_user2016 <- function(id,
+                           width = NULL, height = 300, ratio = c("16by9", "4by3"),
+                           frameborder = 0, allowfullscreen = TRUE){
+
+  id <- c("Events", "useR-international-R-User-conference", "useR2016", id)
+
+  embed_channel9(id, width, height, ratio, frameborder, allowfullscreen)
+}
+
+#' @rdname embed
+#' @export
+#
+embed_user2017 <- function(id,
+                           width = NULL, height = 300, ratio = c("16by9", "4by3"),
+                           frameborder = 0, allowfullscreen = TRUE){
+
+  id <- c(
+    "Events",
+    "useR-international-R-User-conferences",
+    "useR-International-R-User-2017-Conference",
+    id
+  )
+
+  embed_channel9(id, width, height, ratio, frameborder, allowfullscreen)
 }