+++ R/pipe.R
@@ -1,6 +1,6 @@
 #' Pipe functions
 #'
-#' Like dplyr, vembedr also uses the pipe function, \code{\%>\%} to turn
+#' Like dplyr, vembedr also uses the pipe function, `\%>\%` to turn
 #' function composition into a series of imperative statements.
 #'
 #' @importFrom magrittr %>%
@@ -8,6 +8,9 @@
 #' @rdname pipe
 #' @export
 #' @param lhs,rhs An embed object and a function to apply to it
+#'
+#' @keywords internal
+#'
 #' @examples
 #' # Instead of
 #' use_start_time(rickroll_youtube(), "1m35s")