+++ NEWS.md
@@ -1,3 +1,16 @@
+## vembedr 0.1.4
+
+- adds support for [Microsoft Stream](https://www.microsoft.com/en-us/microsoft-365/microsoft-stream), which offers an enterprise video service; this will likely be for use internal to an organization. (#38)
+- if you have a recent version of [usethis](https://usethis.r-lib.org) installed, `suggest_embed()` will copy the suggested code to your clipboard. (#32)
+- sets the default `height` to 300 pixels, and introduces an argument `ratio`, to set the aspect ratio; legal values are `"16by9"` or `"4by3"`. If you set the `width` and `height`, `ratio` is ignored. If you specify only one of `width` or `height`, `ratio` is used to calculate the other. (#33)
+- adds three functions to help with formatting; they are all pipeable with `embed_*()` functions:
+  - `use_rounded()` lets you specify rounded corners. (#25, with @koncina and @ginolhac)
+  - `use_align()` lets you specify a horizontal alignment.
+  - `use_bs_responsive()` if your 'HTML' page includes [Twitter Bootstrap 3](https://getbootstrap.com/docs/3.3/components/#responsive-embed), use this function to make your `<iframe/>` responsive. In this case, the `ratio` is used, `width` and `height` are ignored. (#13)
+  
+- adds functionality for embedding [Box](https://www.box.com) videos, via the `embed_box()` function (#28)
+- version bump for development
+
 ## vembedr 0.1.3
 
 - updates README to show some custom formatting (#25, thanks @koncina and @ginolhac)